split systemd-journal.c
