+++ collectors/systemd-journal.plugin/systemd-units.c
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+#include "systemd-internals.h"
+
